Move alternation methods to new Alterations.cs file
Refactored the Or and Either methods from Rejigs.cs into a new partial class file, Alterations.cs, for better code organization. Renamed the test class and file to AlterationsTests for consistency with naming conventions.
